requirements.txt
array2gif==1.0.4
    # via -r requirements.in
cycler==0.10.0
    # via matplotlib
decorator==4.4.2
    # via networkx
imageio==2.9.0
    # via scikit-image
imutils==0.5.4
    # via -r requirements.in
kiwisolver==1.3.1
    # via matplotlib
matplotlib==3.3.4
    # via
    #   -r requirements.in
    #   scikit-image
networkx==2.5
    # via scikit-image
numpy==1.20.1
    # via
    #   -r requirements.in
    #   array2gif
    #   imageio
    #   matplotlib
    #   opencv-python
    #   pywavelets
    #   scikit-image
    #   scipy
    #   tifffile
opencv-python==4.5.1.48
    # via -r requirements.in
pillow==8.1.0
    # via
    #   imageio
    #   matplotlib
    #   scikit-image
pyparsing==2.4.7
    # via matplotlib
python-dateutil==2.8.1
    # via matplotlib
pywavelets==1.1.1
    # via scikit-image
scikit-image==0.18.1
    # via -r requirements.in
scipy==1.6.0
    # via
    #   -r requirements.in
    #   scikit-image
six==1.15.0
    # via
    #   cycler
    #   python-dateutil
tifffile==2021.2.1
    # via
    #   -r requirements.in
    #   scikit-image